requirements.txt
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
Click==7.0
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3+mkl
opencv-python==4.1.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
PyAutoGUI==0.9.42
pycodestyle==2.5.0
PyGetWindow==0.0.4
Pygments==2.4.0
PyMsgBox==1.0.6
PyQt5==5.11.3
PyQt5-sip==4.19.17
pyqt5-tools==5.11.3.1.4
PyRect==0.1.4
pyrsistent==0.15.2
PyScreeze==0.1.20
python-dateutil==2.8.0
python-dotenv==0.10.2
PyTweening==1.0.3
pytz==2019.1
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.4.4
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
